# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "AMONG API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3065",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/performance/image": {
      "post": {
        "tags": [
          "Performance"
        ],
        "summary": "공연 사진 저장",
        "description": "공연 사진 저장",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/performance/": {
      "post": {
        "tags": [
          "Performance"
        ],
        "summary": "공연 정보 저장",
        "description": "공연 정보 저장",
        "parameters": [
          {
            "name": "performance",
            "in": "body",
            "description": "공연 바디 정보",
            "schema": {
              "$ref": "#/definitions/Performance"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "tags": [
          "Performances"
        ],
        "summary": "모든 공연 정보 보기",
        "description": "모든 공연 정보 보기",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/performance/{SearchWord}/search": {
      "get": {
        "tags": [
          "Performances"
        ],
        "summary": "모든 공연 정보 보기",
        "description": "모든 공연 정보 보기",
        "parameters": [
          {
            "name": "SearchWord",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/performance/id/{id}": {
      "get": {
        "tags": [
          "Performance"
        ],
        "summary": "공연 상세 정보 보기",
        "description": "공연 상세 정보 보기",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/ticket/{id}/detail": {
      "get": {
        "tags": [
          "Ticket"
        ],
        "summary": "티켓 상세 정보 보기",
        "description": "티켓 상세 정보 보기",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/ticket/{ticketId}/cancel": {
      "patch": {
        "tags": [
          "Ticket"
        ],
        "summary": "티켓 판매 등록 취소",
        "description": "티켓 판매 등록 취소",
        "parameters": [
          {
            "name": "ticketId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/ticket/{price}/{ticketId}/register": {
      "patch": {
        "tags": [
          "Ticket"
        ],
        "summary": "티켓 판매 등록",
        "description": "티켓 판매 등록",
        "parameters": [
          {
            "name": "price",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "ticketId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/ticket/{ticketId}/buy": {
      "patch": {
        "tags": [
          "Ticket"
        ],
        "summary": "티켓 구매",
        "description": "티켓 구매",
        "parameters": [
          {
            "name": "ticketId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/ticket/resale": {
      "get": {
        "tags": [
          "Ticket"
        ],
        "summary": "리셀 티켓 보기",
        "description": "리셀 티켓 보기",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/ticket/{ticketId}/useTicket": {
      "patch": {
        "tags": [
          "Ticket"
        ],
        "summary": "티켓 사용",
        "description": "티켓 사용",
        "parameters": [
          {
            "name": "ticketId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/influencer/image": {
      "post": {
        "tags": [
          "Influencer"
        ],
        "summary": "인플루언서 포스터 저장",
        "description": "인플루언서 포스터 저장",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/influencer/register": {
      "post": {
        "tags": [
          "Influencer"
        ],
        "summary": "인플루언서로 변경",
        "description": "인플루언서로 변경",
        "parameters": [
          {
            "name": "Register",
            "in": "body",
            "description": "인플루언서로 변경",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "인플루언서 이름"
                },
                "description": {
                  "type": "string",
                  "example": "소개"
                },
                "image": {
                  "type": "string",
                  "example": "example.png"
                }
              },
              "required": [
                "name",
                "description",
                "image"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/influencer/search": {
      "get": {
        "tags": [
          "Influencer"
        ],
        "summary": "인플루언서들 보기",
        "description": "인플루언서들 보기",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "definitions": {
    "Performance": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "행사 제목 입력"
        },
        "place": {
          "type": "string",
          "example": "행사 장소 입력"
        },
        "time": {
          "type": "string",
          "example": "행사 시간"
        },
        "limitedAge": {
          "type": "string",
          "example": "제한된 나이"
        },
        "term_start_at": {
          "type": "string",
          "example": "YYYY-MM-DD HH:MM:SS"
        },
        "term_end_at": {
          "type": "string",
          "example": "YYYY-MM-DD HH:MM:SS"
        },
        "start_at": {
          "type": "string",
          "example": "YYYY-MM-DD HH:MM:SS"
        },
        "end_at": {
          "type": "string",
          "example": "YYYY-MM-DD HH:MM:SS"
        },
        "description": {
          "type": "string",
          "example": "행사 상세 설명"
        },
        "infos": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "class": {
                "type": "string",
                "example": "좌석 등급"
              },
              "price": {
                "type": "string",
                "example": "좌석 가격"
              },
              "number": {
                "type": "string",
                "example": "좌석 갯수"
              }
            }
          }
        },
        "image": {
          "type": "string",
          "example": "이미지"
        }
      }
    },
...
  }
}